tg-37 vuelta atrás con los cambios de tipos de id, más cambios en algunos tipos

+++ back-end/hospital-api/src/main/java/net/pladema/establishment/repository/entity/InstitutionType.java
@@ -21,7 +21,7 @@ public class InstitutionType {
 	@Id
 	@Column(name = "id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Short id;
 	
 	@Column(name = "description", nullable = false)
 	private String description;
